requirements.txt
Flask==1.0.2
requests==2.22.0
python-dotenv==0.10.5